パスワードリセット関連のディレクトリ構成の変更
